+++ src/module/network/__init__.py
@@ -1,2 +1,1 @@
 from .request_contents import RequestContent, TorrentInfo
-from .notification import PostNotification
