+++ src/module/notification/plugin/__init__.py
@@ -1,0 +1,3 @@
+from .bark import BarkNotification
+from .server_chan import ServerChanNotification
+from .telegram import TelegramNotification
